house_zhkh_core/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
fastapi==0.115.8
h11==0.14.0
idna==3.10
kombu==5.4.2
logstash==0.1.dev0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
utils==1.0.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
